feat : 裁判系统解算没想好怎么搞

+++ EC_lib/bsp/bsp_uart/bsp_usart.h
@@ -35,7 +35,7 @@ typedef struct Usart_Device_{
 		uint8_t rx_buff[USART_RXBUFF_LIMIT];		//单缓冲区使用
 		uint8_t rx_buff2[2][USART_RXBUFF_LIMIT/2];	//双缓存区使用
 	};
-	uint8_t rx_len;									//接收一包数据的大小
+	uint16_t rx_len;								//接收一包数据的大小
 	uint8_t rx_buff_num;							//缓存区数目
     // 接收的回调函数,用于解析接收到的数据
     void (*usart_device_callback)(struct Usart_Device_ *); // callback needs an instance to tell among registered ones
@@ -47,7 +47,7 @@ typedef struct{
 	
 	UART_HandleTypeDef *usart_handle;				// uart句柄
 	
-	uint8_t rx_len;									//接收一包数据的大小
+	uint16_t rx_len;								//接收一包数据的大小
 	uint8_t rx_buff_num;							//缓存区数目
 	//接收回调函数
     void (*usart_device_callback)(Usart_Device_t *);
